// src/app/interfaces/digimon/digimons.interface.ts
// Generated by https://quicktype.io

export interface Digimons {
    content:  ContentDigimons[];
    pageable: Pageable;
}

export interface ContentDigimons {
    id:    number;
    name:  string;
    href:  string;
    image: string;
}

export interface Pageable {
    currentPage:    number;
    elementsOnPage: number;
    totalElements:  number;
    totalPages:     number;
    previousPage:   string;
    nextPage:       string;
}