// src/app/interfaces/digimon/attributes.interface.ts
// Generated by https://quicktype.io

export interface Attributes {
    content:  AttributesContent;
    pageable: Pageable;
}

export interface AttributesContent {
    name:        string;
    description: string;
    fields:      AttributesField[];
}

export interface AttributesField {
    id:   number;
    name: string;
    href: string;
}

export interface Pageable {
    currentPage:    number;
    elementsOnPage: number;
    totalElements:  number;
    totalPages:     number;
    previousPage:   string;
    nextPage:       string;
}